Extract the config of DB
