+++ src/app.module.ts
@@ -3,35 +3,11 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { TodosModule } from './todos/todos.module';
-import { Todo } from './todos/entities/todo.entity';
-import AWS from 'aws-sdk';
-
-const signer = new AWS.RDS.Signer();
-const signerOptions = {
-  region: process.env.AWS_REGION,
-  hostname: process.env.DB_HOST,
-  username: process.env.DB_USERNAME,
-  port: 3306,
-};
+import { AuthModule } from './auth/auth.module';
+import dbConfig from './db.config';
 
 @Module({
-  imports: [
-    TypeOrmModule.forRoot({
-      type: 'mysql',
-      host: process.env.DB_HOST,
-      username: process.env.DB_USERNAME,
-      database: process.env.DB_NAME,
-      entities: [Todo],
-      synchronize: true,
-      ssl: 'Amazon RDS',
-      extra: {
-        authPlugins: {
-          mysql_clear_password: () => () => signer.getAuthToken(signerOptions),
-        },
-      },
-    }),
-    TodosModule,
-  ],
+  imports: [TypeOrmModule.forRoot(dbConfig), TodosModule, AuthModule],
   controllers: [AppController],
   providers: [AppService],
 })
